+++ src/main/java/koval/mobile/myfitnesspal/gui/android/loginPages/LogInPage.java
@@ -8,7 +8,6 @@ import koval.mobile.myfitnesspal.gui.common.downMenuPages.dashboardPage.Dashboar
 import koval.mobile.myfitnesspal.gui.common.loginPages.LogInPageBase;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.support.FindBy;
-import org.openqa.selenium.support.ui.ExpectedConditions;
 
 
 @DeviceType(pageType = DeviceType.Type.ANDROID_PHONE, parentClass = LogInPageBase.class)
@@ -17,8 +16,8 @@ public class LogInPage extends LogInPageBase {
     @FindBy(id = "com.myfitnesspal.android:id/toolbar")
     private TopToolbarModal topToolbarModal;
 
-    @FindBy(xpath = "//android.widget.Button[@text='{L10N:logIn}']")
-    private ExtendedWebElement loginButton;
+    @FindBy(xpath = "//android.widget.Button[@text='%s']")
+    private ExtendedWebElement buttonByText;
 
     @FindBy(xpath = "//*[@resource-id='buttonExistingUserTutorial']/child::*[@class='android.widget.Button']")
     private ExtendedWebElement exitTutorialButton;
@@ -47,24 +46,6 @@ public class LogInPage extends LogInPageBase {
 
 
     @Override
-    public boolean closeNotRespondingPopUpIfPresent() {
-
-        ExtendedWebElement closeAppButton = itemByText.format(CLOSE_APP);
-        boolean isNotRespondingPopUpPresent = closeAppButton.isElementPresent(TIMEOUT_FIFTEEN);
-
-        closeAppButton.clickIfPresent(TIMEOUT_TEN);
-
-
-        return isNotRespondingPopUpPresent;
-    }
-
-    @Override
-    public LogInPageBase closeLoginFailedPopUpIfPresent() {
-        itemByText.format(CLOSE).clickIfPresent();
-        return initPage(getDriver(), LogInPageBase.class);
-    }
-
-    @Override
     public LogInPageBase typePassword(String password) {
         passwordField.type(password, TIMEOUT_FIVE);
         return initPage(getDriver(), LogInPageBase.class);
@@ -73,42 +54,9 @@ public class LogInPage extends LogInPageBase {
 
     @Override
     public DashboardPageBase clickLoginButton() {
-        loginButton.click(TIMEOUT_TEN);
+        buttonByText.format(LOGIN).click(TIMEOUT_TEN);
         return initPage(getDriver(), DashboardPageBase.class);
 
     }
-
-
-    @Override
-    public DashboardPageBase closeNoSubscriptionsPopUpIfPresent() {
-
-        waitUntil(ExpectedConditions.visibilityOfElementLocated(itemByText.format(CANCEL).getBy()), TIMEOUT_FIFTY);
-
-        int attemp = 3;
-        while (itemByText.format(CANCEL).isElementPresent() && attemp > 0) {
-            itemByText.format(CANCEL).click(TIMEOUT_TEN);
-
-            LOGGER.info("[ DASHBOARD PAGE ] Attempt left: {} for clicking on exit No Subscriptions PopUp button", attemp);
-            attemp--;
-        }
-
-        return initPage(getDriver(), DashboardPageBase.class);
-    }
-
-
-    @Override
-    public void closeUserTutorialBoxIfPresent() {
-
-        int attemp = 3;
-        while (exitTutorialButton.isElementPresent(TIMEOUT_TEN) && attemp > 0) {
-            exitTutorialButton.click(TIMEOUT_TEN);
-
-            LOGGER.info("[ DASHBOARD PAGE ] Attempt left: {} for clicking on exit tutorial button", attemp);
-            attemp--;
-        }
-
-        initPage(getDriver(), DashboardPageBase.class);
-    }
-
 
 }
