+++ src/main/java/koval/mobile/myfitnesspal/gui/android/loginPages/WelcomePage.java
@@ -1,12 +1,15 @@
 package koval.mobile.myfitnesspal.gui.android.loginPages;
 
 
+import com.qaprosoft.carina.core.gui.AbstractPage;
 import com.zebrunner.carina.utils.factory.DeviceType;
-import com.zebrunner.carina.utils.resources.L10N;
 import koval.mobile.myfitnesspal.gui.common.loginPages.LogInPageBase;
 import koval.mobile.myfitnesspal.gui.common.loginPages.SignUpPageBase;
 import koval.mobile.myfitnesspal.gui.common.loginPages.WelcomePageBase;
 import org.openqa.selenium.WebDriver;
+
+
+
 
 @DeviceType(pageType = DeviceType.Type.ANDROID_PHONE, parentClass = WelcomePageBase.class)
 public class WelcomePage extends WelcomePageBase {
@@ -26,11 +29,34 @@ public class WelcomePage extends WelcomePageBase {
         return initPage(getDriver(), SignUpPageBase.class);
     }
 
+//    public static String getText(String key) {
+//        // Determine the current language or locale
+//        Locale currentLocale = Locale.forLanguageTag("en_US"); // Implement this method to get the current locale
+//
+//        LOGGER.info("cur locale: " + currentLocale);
+//        // Load the appropriate resource bundle based on the current locale
+//        ResourceBundle bundle = ResourceBundle.getBundle("locale", currentLocale);
+//
+//        // Retrieve the localized text for the key
+//
+//        return bundle.getString(key);
+//    }
+
+
+
+
     @Override
     public LogInPageBase logIn() {
         itemByText.format(LOGIN).click(TIMEOUT_TEN);
+
         return initPage(getDriver(), LogInPageBase.class);
     }
 
+    @Override
+    public AbstractPage clickButton(String buttonText) {
+
+        itemByText.format(buttonText).click(TIMEOUT_TEN);
+        return initPage(getDriver(), LogInPageBase.class);
+    }
 
 }
