+++ src/main/java/koval/mobile/myfitnesspal/gui/common/loginPages/WelcomePageBase.java
@@ -1,5 +1,6 @@
 package koval.mobile.myfitnesspal.gui.common.loginPages;
 
+import com.qaprosoft.carina.core.gui.AbstractPage;
 import koval.mobile.myfitnesspal.gui.MyAbstractPageBase;
 import org.openqa.selenium.WebDriver;
 
@@ -15,4 +16,6 @@ public abstract class WelcomePageBase extends MyAbstractPageBase {
     public abstract SignUpPageBase signUp();
 
     public abstract LogInPageBase logIn();
+
+    public abstract AbstractPage clickButton(String buttonText);
 }
