+++ src/test/java/koval/mobile/myfitnesspal/signUp/FitnessPalSignUpTest.java
@@ -1,0 +1,33 @@
+package koval.mobile.myfitnesspal.signUp;
+
+import com.zebrunner.agent.core.annotation.TestLabel;
+import com.zebrunner.carina.core.registrar.ownership.MethodOwner;
+import koval.mobile.myfitnesspal.gui.common.downMenuPages.dashboardPage.DashboardPageBase;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.testng.Assert;
+import org.testng.annotations.Test;
+
+import java.lang.invoke.MethodHandles;
+
+
+public class FitnessPalSignUpTest extends SignUpTest {
+
+    private static final Logger LOGGER = LoggerFactory.getLogger(MethodHandles.lookup().lookupClass());
+
+
+    @Test()
+    @MethodOwner(owner = "koval")
+    @TestLabel(name = "feature", value = {"mobile", "regression"})
+    public void signUpSimpleUserTest() {
+
+        DashboardPageBase dashboardPageBase = initPage(getDriver(), DashboardPageBase.class);
+        dashboardPageBase.lastMealPopUpSkip();
+
+        Assert.assertTrue(dashboardPageBase.isPageOpened(), "[ DASHBOARD PAGE ] Dashboard page is not opened!");
+
+        Assert.assertFalse(dashboardPageBase.isUserPremium(),"[ DASHBOARD PAGE ] User is premium!" );
+    }
+
+
+}
