+++ src/test/resources/locale.properties
@@ -1,0 +1,64 @@
+allow=Allow
+logIn=Log In
+signUp=Sign Up
+email_address=Email Address
+password=Password
+cancel=Cancel
+close=Close
+calories=Calories
+go_premium=Go Premium
+diary=Diary
+dashboard=Dashboard
+newsfeed=Newsfeed
+plans=Plans
+more=More
+don't_ask_me_again=Don''t ask me again
+got_it=Got it
+breakfast=Breakfast
+lunch=Lunch
+dinner=Dinner
+snacks=Snacks
+connect_a_step_tracker=Connect a step tracker
+exercise=Exercise
+all=All
+my_meals=My Meals
+my_recipes=My Recipes
+my_foods=My Foods
+create_a_meal=Create a Meal
+copy_a_meal=Copy Previous Meal
+create_a_recipe=Create a Recipe
+create_a_food=Create a Food
+no_Thanks=No Thanks
+no_thanks_nutrient=No Thanks
+no_THANKS=No Thanks
+no_thanks=No thanks
+save=Save
+next=Next
+add_to_diary=Add to Diary
+add=Add
+accept=Accept
+new_recipe=New Recipe
+home=Home
+close_app=Close app
+search_for_a_food=Search for a food
+food=Food
+add_exercise_caps=ADD EXERCISE
+delete=Delete
+edit=Edit
+more_options=More options
+cardiovascular=Cardiovascular
+strength=Strength
+water=Water
+DiaryPage.title=Diary
+my_info=My Info
+my_items=My Items
+my_posts=My Posts
+meals=Meals
+recipes=Recipes
+foods=Foods
+cardio=Cardio
+yes=Yes
+log_out=Log Out
+net_crabs=Net Carbs
+
+
