+++ src/test/resources/locale_es_ES.properties
@@ -1,0 +1,62 @@
+allow=Permitir
+logIn=Iniciar sesión
+signUp=Registrarse
+email_address=Dirección de email
+password=Contraseña
+cancel=Cancelar
+close=Cerrar
+calories=Calorías
+go_premium=Pasar a Premium
+diary=Diario
+dashboard=Tablero
+newsfeed=Noticias
+plans=Planes
+more=Más
+don't_ask_me_again=No volver a preguntar
+got_it=Entendido
+breakfast=Desayuno
+lunch=Almuerzo
+dinner=Cena
+snacks=Aperitivos
+connect_a_step_tracker=Conecta un monitor de pasos
+exercise=Ejercicio
+all=Todos
+my_meals=Mis comidas
+my_recipes=Mis recetas
+my_foods=Mis alimentos
+create_a_meal=Crea una comida
+copy_a_meal=Copiar comida anterior
+create_a_recipe=Crea una receta
+create_a_food=Crear un alimento
+no_thanks=No, gracias
+no_thanks_nutrient=No, gracias
+no_THANKS=NO, GRACIAS
+no_Thanks=No, Gracias
+close_app=Cerrar aplicación
+save=Guardar
+next=Siguiente
+add_to_diary=Añadir al diario
+add=Añadir
+edit=Editar
+more_options=Más opciones
+accept=Aceptar
+new_recipe=Nueva receta
+home=Inicio
+search_for_a_food=Busca un alimento
+food=alimento
+add_exercise_caps=AÑADIR EJERCICIO
+delete=Borrar
+cardiovascular=Aeróbico
+strength=Musculación
+water=Agua
+DiaryPage.title=Diario
+my_info=Mi información
+my_items=Mis artículos
+my_posts=Mis posts
+meals=Comidas
+recipes=Recetas
+foods=Alimentos
+cardio=aeróbicos
+yes=Sí
+log_out=Cerrar sesión
+net_crabs=Carbohidratos netos
