+++ src/test/resources/testng_suites/koval/carinademo/android_MenuTest.xml
@@ -7,7 +7,7 @@
 
     <test name="Left menu check test">
         <classes>
-            <class name="koval.carinademo.LeftMenuCheckTest">
+            <class name="koval.mobile.carinademo.LeftMenuCheckTest">
 <!--                <methods>-->
 <!--                    <include name=""/>-->
 <!--                    <include name=""/>-->
