+++ src/test/resources/testng_suites/koval/myfit/GoogleFitTest.xml
@@ -7,7 +7,7 @@
 
     <test name="FitnessPal">
         <classes>
-            <class name="koval.myfitnesspal.login.FitnessPalTest">
+            <class name="koval.mobile.myfitnesspal.login.FitnessPalTest">
             </class>
         </classes>
     </test>
